refactor: 에러 코드 각 도메인으로 이동 #131
